+++ database/migrations/2018_03_26_090844_add_returnable_enabled_to_coupon_types.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddReturnableEnabledToCouponTypes extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('coupon_types', function (Blueprint $table) {
+            $table->boolean('returnable')->default(true)->after('order');
+            $table->boolean('enabled')->default(true)->after('order');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('coupon_types', function (Blueprint $table) {
+            $table->dropColumn('enabled');
+            $table->dropColumn('returnable');
+        });
+    }
+}